Move Settings to dedicated classes
